Feature: Sliders will accept input even if the user moves the mouse away
I'm not happy with calling the mousePressed() function and dumping a list of all the sliders into it. A better way would be to make an array of functions to call on mousePressed() and have every slider add itself to the list on creation. Since we aren't likely to add many more sliders it's probably fine.